Fix circular dependencies
Split code to fix circular dependencies
