+++ specifyweb/frontend/js_src/lib/components/Atoms/Internationalization.ts
@@ -7,6 +7,16 @@ import type { LocalizedString } from 'typesafe-i18n';
 import { LANGUAGE } from '../../localization/utils/config';
 import type { RA } from '../../utils/types';
 import { capitalize } from '../../utils/utils';
+import {
+  DAY,
+  HOUR,
+  MILLISECONDS,
+  MINUTE,
+  MONTH,
+  SECOND,
+  WEEK,
+  YEAR,
+} from './timeUnits';
 
 /* This is an incomplete definition. For complete, see MDN Docs */
 // eslint-disable-next-line @typescript-eslint/no-namespace
@@ -126,16 +136,6 @@ const numberFormatter = new Intl.NumberFormat(LANGUAGE);
 export const formatNumber = (number: number): LocalizedString =>
   numberFormatter.format(number) as LocalizedString;
 
-/* eslint-disable @typescript-eslint/no-magic-numbers */
-export const MILLISECONDS = 1000;
-const SECOND = 1;
-const MINUTE = 60 * SECOND;
-const HOUR = 60 * MINUTE;
-export const DAY = 24 * HOUR;
-const WEEK = 7 * DAY;
-const MONTH = 4 * WEEK;
-const YEAR = 12 * MONTH;
-/* eslint-enable @typescript-eslint/no-magic-numbers */
 const relativeDate = new Intl.RelativeTimeFormat(LANGUAGE, {
   numeric: 'auto',
   style: 'long',
