+++ specifyweb/frontend/js_src/lib/components/Preferences/BasePreferences.tsx
@@ -10,7 +10,6 @@ import { parseValue } from '../../utils/parser/parse';
 import type { GetOrSet, RA } from '../../utils/types';
 import { filterArray, setDevelopmentGlobal } from '../../utils/types';
 import { keysToLowerCase, replaceKey } from '../../utils/utils';
-import { MILLISECONDS } from '../Atoms/Internationalization';
 import { softFail } from '../Errors/Crash';
 import {
   cachableUrl,
@@ -19,6 +18,7 @@ import {
 } from '../InitialContext';
 import { formatUrl } from '../Router/queryString';
 import { GenericPreferences, PreferenceItem } from './types';
+import { MILLISECONDS } from '../Atoms/timeUnits';
 
 /* eslint-disable functional/no-this-expression */
 // TESTS: add tests for user preferences
